PostgreTest

+++ anyline-simple-data-jdbc-dialect/anyline-simple-data-jdbc-postgresql/src/test/java/org/anyline/simple/postgre/PostgreTest.java
@@ -100,7 +100,7 @@ public class PostgreTest {
             row.put("NAME", "N"+i);
             set.add(row);
         }
-        int qty = service.insert(table, set);
+        long qty = service.insert(table, set);
         log.warn(LogUtil.format("[批量插入][影响行数:{}][生成主键:{}]", 36), qty, set.getStrings("ID"));
         Assertions.assertEquals(qty , 9);
